pyproject.toml
[project]
name = "virtualizarr-data-pipelines"
version = "0.1.0"
description = "Virtualizarr/Icechunk ingestion infrastructure"
authors = [{name = "Sean Harkins", email = "sean@developmentseed.org"}]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aws-cdk-lib",
    "constructs>=10.0.0",
]

[tool.uv.sources]
virtualizarr-processor = { path = "lambda/virtualizarr-processor", editable = true}

[dependency-groups]
dev = [
  "aws-lambda-powertools>=2.30.0",
  "aws-lambda-typing>=2.20.0",
  "aws-xray-sdk",
  "boto3-stubs[batch,s3,sqs]>=0.37.36",
  "mypy",
  "pre-commit",
  "nodeenv",
  "pydantic-settings>=2.8.1",
  "pytest",
  "ruff",
  "virtualizarr-processor",
]


[tool.ruff]
line-length = 88
target-version = "py311"
exclude = ["cdk.out", ".venv", "node_modules"]

[tool.ruff.lint]
select = ["E", "F", "I"]
per-file-ignores = {}

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
explicit_package_bases = true
files = ["lambda", "cdk"]
